refactor: 统一并修正图标处理模块引用路径
将社区插件图标处理器和相关工具的导入路径调整为统一的 util/ 目录，
并在主入口文件中使用新的处理器位置。主要改动：
- CommunityPluginIconHandler.ts 中将 addIconToPluginNavItem 和
  AbstractIconHandler 的导入路径改为 @src/util 下的模块，修复了
  不一致的相对路径。
- src/main.ts 中替换 CommunityPluginIconHandler 的导入来源为
  ./service/CommunityPluginIconHandler，并移除旧的 ./util/handlers 路径。
- 在 main.ts 中提前引入 CommunityPluginIconHandler，确保在插件
  初始化阶段使用正确的处理器实现。

这样做是为了统一模块组织、减少相对路径混淆并避免运行时导入错误，
提高代码可维护性和模块定位的一致性。
